# perf(es/minifier): Box `VarUsageInfo` of `ProgramData`