--- a/.changeset/late-maps-jump.md
+++ b/.changeset/late-maps-jump.md
@@ -1,0 +1,6 @@
+---
+swc_core: minor
+swc_ecma_minifier: minor
+---
+
+perf(es/minifier): Reduce memory allocations