@kdy1 2025-01-19
Original: 
```
heap stats:     peak       total       freed     current        unit       count
  reserved:     2.0 GiB     2.0 GiB     0           2.0 GiB
 committed:     1.9 GiB     2.0 GiB   736.4 MiB     1.2 GiB
     reset:     0
    purged:   501.8 MiB
   touched:    64.2 KiB   321.2 KiB   684.6 MiB  -684.3 MiB                          ok
  segments:     4           5           2           3                                not all freed
-abandoned:     0           0           0           0                                ok
   -cached:     0           0           0           0                                ok
     pages:     0           0           5.9 Ki     -5.9 Ki                           ok
-abandoned:     0           0           0           0                                ok
 -extended:     0
 -noretire:     0
    arenas:     2
-crossover:     0
 -rollback:     0
     mmaps:     0
   commits:     0
    resets:     0
    purges:   275
   threads:    31          31           0          31                                not all freed
  searches:     0.0 avg
numa nodes:     1
   elapsed:     2.648 s
   process: user: 3.060 s, system: 0.575 s, faults: 0, rss: 159.0 MiB, commit: 1.9 GiB
```